riscv-regs: Add hpmcounter CSR definitions
All 32 hpmcounter registers share the same layout and we'd like to
index them at runtime for PMU virtualization, so put them in an array
of RiscvCsrInterface trait objects. Some code duplication is required
to actually instantiate the registers, however this could potentially
be addressed with a procedural macro. That's left as a TODO for now
given that hpmcounter is the only array of CSRs that we currently have
to deal with.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>