Vm: Separate the vCPU context switch from the virtualization mode switch
Up until now we've been saving/restoring (almost) everything on enter/exit
from virtualization, but with PMU support on the horizon, this approach is
going to make every VM enter/exit pretty expensive.

The two "switches" are now:
 - the vCPU context switch, preformed by VmCpu::activate() (from parent
   to child) and ActiveVmCpu::drop() (from child back to parent) for
   switching out state shared between vCPUs (e.g. VS-level CSRs).
 - the V mode switch, performed by ActiveVmCpu::run_to_exit() for
   switching out state shared between the vCPU and Salus in HS-mode
   (e.g. GPRs, FPU, some CSRs).

The latter remains much the same as before, except that we've moved out
save/restore of the VS-level CSRs. The former is implemented by saving
an exclusive reference to the parent ActiveVmCpu when activating a child
VmCpu: the construction of the child ActiveVmCpu saves the parent's state
and restores the child's state, while the destruction of the child
ActiveVmCpu saves the child's state and restores the parent's.

Finally, to take advantage of the separate context vs mode switch, make
vCPU activation a one-time operation at the start of the run loop in
run_vcpu() rather than doing it on each iteration.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>