pci: Verify VM-assigned BAR addresses
While assigning an unowned address to a BAR or bridge window won't give
a VM access to that unowned memory, it could potentially create an alias
with a PCI resource for a device not owned by the VM (for example the
IOMMU, which will be owned by Salus and hidden from all VMs). Prevent
this situation by intercepting attempts by the VM to enable memory or IO
space and verifying that all the device's BARs and bridge windows map to
physical address ranges owned by that VM. If a BAR or bridge window
assignment maps unowned memory, the write to the corresponding enable bit
in COMMAND is discarded.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>