page_table: remove tracker form `PageTableInner`
For Sv48 page tables that will provide hypervisor-virtual addressing, no
interaction with page tracking is possible or desired. So break the
dependency that all page tables reflect page ownership when that's only
appropriate for moving pages to/from VMs.

Instead move the tracking and ownership logic to the `PlatformPageTable`
level, which is 1:1 with `PageTableInner`. This change doesn't affect
the ownership of pages, just where it is tracked so that non-owning (HS
and HU mode) page tables can share the common parts of `PageTableInner`.

This moves the `Drop` implementation logic up a layer to
`PlatformPageTable` as the page tracker releases are specific to VM page
assignements and not wanted when dropping an Sv48.

Signed-off-by: Dylan Reid <dgreid@rivosinc.com>