pci: Support claiming devices for hypervisor-internal usage
In order to support the IOMMU as a PCI device, add limited support for
enabling the use of PCI devices from within Salus. Devices must be on
the root bus (i.e. not behind any bridges) so that we don't have to deal
with sizing bridge windows. The BAR resources for such devices are
allocated from the end of the host bridge windows.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>