iommu: Add support for building a device directory table
The device directory table (DDT) is used by the IOMMU to map from device
ID to translation context. For our purposes we assume that the device ID
is the PCI requester ID.

Each device context configures 1st-stage, 2nd-stage, and MSI translation
for the device. For now we are simply passing through PCI devices with
no virtualized IOMMU support in guest VMs, so we need only deal with
2nd-stage and MSI translation (i.e. first-stage translation is "bare").

Populate the DDT when the IOMMU is probed with (invalid) device contexts
for each present PCI device. We'll enable translation for these devices
in follow-on patches.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>