iommu: Add support for building MSI page tables
MSI page tables are a flat (single-level) paging structure used to provide
translation for MSI writes. The input to an MSI page table is the guest
physical address of an MSI write (i.e. the address of an virtualized IMSIC
in guest physical address space) and the output is the supervisor physical
address of an IMSIC interrupt file (or a memory-resident interrupt file,
which isn't yet supported). The recently-introduced ImsicGeometry and
ImsicLocation to specify mappings in the MSI page table.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>