iommu: Add API for attaching/detaching devices
Add attach_pci_device() and detach_pci_device() which enable and disable
DMA translation for a PCI device. A reference count is held to the device
owner's GSCID for the lifetime of the attachment to prevent the use of
stale GSCIDs.

Also add a fence() API issuing IOTINVAL.GVMA commands which are necessary
after 2nd-stage page table or MSI page table updates.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>