riscv-page-tables: Rename `PlatformPageTable` and `GuestStagePageTable`
Much of what `PlatfromPageTable` does is too specific to stage-2 VM
paging for it to be useful for Sv48 or Sv39. Accept that and rename it.
This will allow for a new type to be introduced to handle first stage
translations in a later commit.

To clear the way for the new `GuestStagePageTable` rename the PagingMode
shortcuts to better indicate they are paging modes, not page tables.
This allows for consistent naming of both the paging modes and the
implemented page tables.

Signed-off-by: Dylan Reid <dgreid@rivosinc.com>