riscv-page-tables: Add `FirstStagePageTable` type.
Unlike the page tables used for VMs. Pages tables for HS mode don't
track the ownership of the pages assigned to them. Since page ownership
decisions are made by the HS mode hypervisor, it gets to map all of
memory without having to the entity that owns it. Many pages mapped to
HS mode by Sv48 will also be mapped to guests or the host.

To enable this, allow unsafely mapping addresses (not pages) to the
host.

Signed-off-by: Dylan Reid <dgreid@rivosinc.com>