page-tracking: Add helper to reserve and take a region
For building Sv48 PTEs, main will want to reserve a chunk of memory from
the map early in boot and use that memory as PTE pages. Add a helper
that marks a region as reserved and returns a `SequentialPages` that can
be used to consume the pages that were reserved.

Signed-off-by: Dylan Reid <dgreid@rivosinc.com>