main: Enable sv48 paging in HS mode
Turns on paging in salus.

This will be useful for mapping different regions with different
permissions. It is also a pre-requisite for running tasks in u-mode.

There are two big hacks for allowing writes to the UART and reset device
in qemu as salus currently hard codes those device addresses.

Another caveat is that this is slow. the tellus test time went from 25
to 36 seconds on my MBP. Hopefully this will improve with huge page
support, otherwise Salus will have to handle page faults and populate on
demand as most memory won't be touched.

Signed-off-by: Dylan Reid <dgreid@rivosinc.com>