attestation: PCR mapping fix and update
This is an update for the current PCR mapping. The measuremenents we're
including are the first 2 PCRs from the RoT, the TVM pages and
configuration and any dynamic measurement the TVM decides to make.
This also includes an implicit fix as we were including the static OS
measurements as part of the TVM TCB.

Signed-off-by: Samuel Ortiz <sameo@rivosinc.com>