sbi: Define vCPU shared-memory state interface
In order to implement more efficient handling of TVM vCPU exits, define
an interface for using a shared-memory area to communicate the exit
status of a vCPU between the TSM and the host.

The layout of the shared-memory state area is defined by the TSM, and can
be enumerated by the host using the TvmCpuGetMemLayout call. The base
address of the page(s) used for a vCPU's shared-memory state area are then
registered on the call to TvmCpuCreate.

This is intended to replace the TvmCpu{Get,Set}Register calls, which will
be removed once the shared-memory communication is implemented.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>