VmPages: Make VmPagesMapper type-safe
Currently, once a VmPagesMapper is constructed, one can use any sort of
page with it regardless of the region type it was constructed in. Make
it type-safe by adding a type parameter indicating which type of mapper
it is and by having map_page() for that type only support the proper
PhysPage.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>