Vm: Use wrapper types to enforce API separation
Similar to VmPages, remove the state type argument from Vm and introduce
a VmRef type that serves as a reference to a Vm in a particular state.
The various VmRef types are then used to implement the appropriate
functionality for a VM in that state. This immediately enables a couple
of small cleanups: finalize() no longer needs to consume the Vm struct
and we can now implement drop() for Vm.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>