Rework vCPU shared memory register set enumeration
Instead of having single call that writes the layout to host memory and
that the host must retry if they provide insufficient buffer space, take
an approach similar to the SBI PMU interface where a one call returns
the number of register sets (TvmCpuNumRegisterSets) and another returns
the location of a specific register set (TvmCpuGetRegisterSet).

It is expected that the host will call TvmCpuGetRegisterSet for each index
in the range 0 to the value returned by TvmCpuNumRegisterSets in order to
fully enumerate the layout of the vCPU shared-memory state area. Using
an intermediate index (vs iterating over RegisterSetIds) was chosen in
order to help the host discover potential compatibility issues (e.g.
if the TSM exposes a register set the host isn't expecting or aware of).

As part of this change, RegisterSetLocation is compressed to a single u32
so that it's guaranteed to fit in a single general-purpose register. The
version field is dropped and instead the versioning is implicit from the
RegisterSetId.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>