sbi: Define calls for binding/unbinding vCPUs to IMSIC files
In order for a TVM vCPU to use hardware AIA virtualization -- that is,
to use an IMSIC guest interrupt file to enable delivery of external
interrupts directly to VS level -- we need to give the host the ability
to assign (bind) or unassign (unbind) TVM vCPUs to/from these guest
interrupt files.

The new calls are:
- tvm_cpu_bind_imsic() to bind a vCPU to a confidential guest interrupt
  file
- tvm_cpu_unbind_imsic_begin() to invalidate the mapping for a vCPU's IMSIC
- tvm_cpu_unbind_imsic_end() to complete the unbind process and save the
  vCPU's interrupt state to a software interrupt file

The host is reponsible for performing a TVM TLB invalidation sequence
between unbind_imsic_begin() and unbind_imsic_end(). The process for
saving and restoring interrupt file state otherwise follows the one
outlined in sections 7.1.2 and 9.3.3 of the AIA specification.

Note that binding a vCPU to a guest interrupt file also binds the vCPU
to that physical CPU. As a result, tvm_cpu_run() will fail if the vCPU
is unbound or it is called on a different physical CPU from which the
vCPU is bound.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>