Add 'doc' Makefile target
Add a target to the Makefile for generating docs. Run that step in CI to
make sure it works.

test_workloads is causing a rustdoc panic, so exclude it from docs.
Unblocking the rest of the docs is more important than fixing that.
Remove it from the salus deps in Cargo.toml as salus doesn't depend on
the workloads it runs, they are built by different Makefile targets
anyways.

Automatically uploading to github pages would be great but doesn't seem
to have a clean solution yet. For now, being able to generate and view
the docs locally by opening the output under 'target/' is still a big
improvement.

Open and browse/search from:
target/riscv64gc-unknown-none-elf/doc/salus/index.html

Signed-off-by: Dylan Reid <dgreid@rivosinc.com>

sqme