imsic: Add support for saving interrupt state to SW interrupt files
When swapping out or migrating vCPUs we will need to save the interrupt
state of the vCPU to a SW interrupt file. Add a SwFile structure to
hold this state and add support to the Imsic driver to save/restore
guest interrupt file state to a SwFile. This procedure is described in
more detail in section 9.3 of the AIA specification. Note that SwFile
isn't intended to serve as an MRIF just yet -- there's no IOMMU support
for them presently -- but it's structure closely resembles an MRIF so
the transition to an actual MRIF should be straightforward.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>