Remove hard-wired page numbers.
A few values dependent on structure size were hardwired to explicit
numbers. As these structures will change in the future, calculate them
at build time.

TVM_STATE_PAGES was depending on the size of the PageArc containing
GuestVmInner. Being GuestVmInner a generic (for type T), the value has
been moved to a constant method of GuestVm.

As a consequence, TVM_STATE_PAGES is also moved to a VmPages constant
method.