vCPU: Allocate structure on VCPU creation.
This change moves the need to donate memory required to hold vCPU
state from creation of the VM to creation of the vCPU.

Both TvmCreate and TvmCpuCreate interfaces are modified as a result.

VmCpus is now a fixed array of `Once` elements.  A `PageBox`
containing the whole vCPU structure is created when a vCPU is added to
the guest.

One of the victims of this refactor is the `VmCpus::num_cpus()`. This
was used both as the number of vCPUs presents and the maximum vCPU ID
present. Now that vCPUs are not necessarily created in order this
function need to be redefined in future works.