page-tracker: Add page states for runtime unassignment
We will need support runtime unassignment of pages for IMSIC bind/unbind,
as well as dynamic memory (un)sharing. Add an "Unassigning" state which
represents a page that has been invalidated at a particular TLB version
in a VM's page table, but hasn't yet completed unassignment. Once
unassignment is completed, the page is returned to the previous owner in
the "Converted" state.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>