main: Correct host VM physical address space size calculation
We were using the delta between the start of DRAM and end of the memory
map to calculate the size of the host VM's guest physical address space,
but this fails to account for the MMIO regions (IMSIC and 32-bit PCI BAR)
that lie below the start of DRAM and also end up getting mapped into the
host VM's address space. Use the delta between the start and end of the
physical memory map instead.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>