Remove page_type from TsmConvert/Reclaim pages
The page_type parameter in the TsmConvertPages and TsmReclaimPages was
hardcoded to PageSize4K for all intents and purposes. The consensus
following an internal discussion about the APIs was that we should
remove the page_type parameter in APIs that don't directly affect
page tables.

The rationale behind the change is that the HW mechanism for protecting
confidential memory is not necessarily page-table based (example: PMP).
Furthermore, even for page-table based implementations, there may be a
mismatch between the requested page size and what the implementation
can reasonably provide, resulting in additional complexity in both the
host and TSM. Note that the TSM implementation is still free to
internally optimize conversion of greater than 4kB-sized ranges if HW
support is available.