Vm: Implement IMSIC bind/unbind calls
The bind operation restores interrupt state from a SW interrupt file
to the selected physical guest interrupt file and maps the guest
interrupt file page into the guest's address space.

The unbind operation saves interrupt state to a SW interrupt file and
unmaps the guest interrupt file page from the guest's address space.
Unbind is split into two phases as the host is required to perform
a TLB invalidation sequence for the guest.

Running a vCPU in a TVM with AIA support enabled now requires that
the vCPU be bound to an interrupt file on the same physical CPU.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>