sbi: Add debug console extension
The new debug console extension seems stable. We will give it a try in
salus and see if it helps boot time.

Note than until the supported SBI version is updated, linux will ignore
this extension. Tested with Tellus and with the version check hacked out
of the kernel.

The legacy SBI console is moved to SbiConsoleV01. Salus continues to use
this extension so that it works with the current OpenSBI
implementation.

Signed-off-by: Dylan Reid <dgreid@rivosinc.com>