VmCpu: Reorganize locking
In preparation for interrupt injection -- where we'll need access to the
external interrupt of a running vCPU -- rework the locking within VmCpu
by separating it into three separate locks:
 - VmCpu::status for powered/runnable status flag
 - VmCpu::arch for registers and other (local) architectural state
 - VmCpu::ext_interrupts for external interrupt / IMSIC state

VmCpu::status is checked when powering on or running (activating) a vCPU
as before, while VmCpu::arch is held while the vCPU is activated so that
we can access the vCPU's registers, TLB version, etc.

VmCpu::ext_interrupts is taken when binding/unbinding an IMSIC and will
be used to guard interrupt injection / allowlisting in follow-on patches.

Also rename some of the inner structures to be more descriptive and
avoid overloading "state" too much.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>