VmPages: Fix active TLB version check
We presently use the current TLB version -- which is not necessarily the
minimum active version -- when getting references to converted or unassigned
pages. Oops. Define min_version() which returns the previous TLB version if
there are active references and use it when acquiring references to pages
which have been invalidated.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>