VmPages: Protect VmRegionList with a RwLock
The list of confidential vs shared regions will soon become dynamic post
VM finalization, so map operations on a running VM will need to protect
against concurrent modifications of the region list. Make the VmRegionList
lock a RwLock and hold in read mode while mapping pages in VmPagesMapper.
This also requires moving the lock out of VmRegionList, which is something
we'll need anyway to support dynamic (un)sharing of memory.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>