VmPages: Implement basic support for memory (un)sharing
Add support for {Share,Unshare}Memory calls from the guest where the
region to be converted is unpopulated. This is handled by converting
the specified chunk of address space in VmRegionList to the desired
type, assuming the range falls completely within a region of the
expected source type (e.g. confidential for confidential -> shared
conversion).

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>