riscv-page-tables: Make unassign_range() more generic
Simlar to the change to invalidate_range(), make unassign_range() take
a closure to check if the PTE can be cleared rather than checking
PageTracker directly in order to support the unassignment of shared
pages. Also, rename it to unmap_range() since it's now clearing PTEs
and not touching page states directly.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>