riscv-page-tables: Introduce get_mapped_pages()
This returns an iterator over a mapped range of virtual address space
if a predicate function returns true for the entire range. Use it to
acquire references to shared pages without having to put them into a
PageList. We'll drop the need to acquire an exclusive reference to
shared pages entirely in a follow-on patch.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>