page-tracker: Remove need to lock shared pages
"Shared" pages should be able to be shared by the host however the host
wants and so attempts to concurrently acquire a reference to a shared
page shouldn't cause errors. We're now synchronizing access to shared
pages externally from PageTracker (using the PageTable lock) and thus
no longer need the "lock" bit in each PageInfo. Remove it, and go back
to an explicit "ConvertedLocked" state for the pages in the process of
assingment/reclaim since it is still an error to try to concurrently
assign/reclaim a confidential page.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>