VmPages: Support (un)sharing of populated memory regions
Invalidate any mapped pages in the region to be (un)shared and mark the
region as in the process of conversion. Complete any pending operations
when the last reference to an old TLB version is dropped by walking the
VmRegionList and unmapping any regions that were converted at now-stale
TLB versions. If the range is populated, the calling vCPU is blocked
until its host completes a TLB fence operation.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>