Use page allocator to allocate hypervisor PTEs.
Allocate the PTEs with the page allocator rather than by reserving
memory pages.

This also fixes an issue in current code where we were reserving
pages *after* we set up paging for it (this happened when creating
the page allocator, that happened after paging setup.

Also make the code clear on when we're not allowed to modify the map
anymore, and enforce that at compiler level by removing mutability of
the hardware memory map.

Signed-off-by: Gianluca Guida <gianluca@rivosinc.com>