VmCpu: HostCpuState -> HypervisorCpuState
The term "host" is being overloaded, referring to both the host of a TVM
(which itself is a VS-level guest) and Salus itself. Lets refer to the
latter as "hypervisor" and rename HostCpuState to HypervisorCpuState.
This is in preparation for needing to access the host (as in, the host
of the guest TVM) state from VmCpu.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>