VmCpu: Make parent_vcpu a more generic host_context
In order make supporting vCPU's which are hosted by something other than
another vCPU less awkward, such as within Salus itself to host the host
VM or, in the future, a mode of operation where the host shares HS mode
with Salus, rename parent_vcpu to host_context and VmCpuSaveState to
HostCpuContext to make them a more generic interface to the hosting
(v)CPU's state.

We'll use this in a follow-on patch to unify exit-reporting between
guest TVMs (to the host VM via shared memory) and the host VM (to
the Salus-internal run loop via internal structures).

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>