sbi: Report SCAUSE/STVAL values via CSR rather than shared memory
SCAUSE and STVAL can be reported to the host via CSR regardless of the
deployment model. When the host is itself a VM (i.e. the model Salus
currently supports), VSCAUSE/VSTVAL can be written to report the values
to the host. When the host is in HS mode, then SCAUSE/STVAL can of
course be used directly.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>