sbi: Use NACL-like structure for TsmShmemArea
Change the layout of the TsmShmemArea struct to match that of the proposed
nested acceleration ("NACL") extension[0]. This layout reserves enough
space for all possible HS and VS CSRs, future-proofing us against possible
future extensions.

Note that although we're using the same structure layout as the NACL
extension, we're defining the contents of the scratch space differently
and the way the TEE-Host extension interacts with the CSR area is
different.

[0] https://github.com/avpatel/linux/commits/riscv_sbi_nested_v1

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>