Introduce riscv-elf
riscv-elf is a library to parse an elf file in memory in Rust. It is
esplicitly tailored to salus and riscv. Work has been done (mainly
through fuzzing) to check that the library can handle malformed ELF
functions.

It is a simple library focused on loading an elf. Its main use is to
return an `ElfSegment` iterator that mirrors the Program Header
section and its data content.

Signed-off-by: Gianluca Guida <gianluca@rivosinc.com>