VmCpu: Emulate HIE/HIP and HGEIE/HGEIP for the host VM
In order to support injecting supervisor guest external interrupts into
the host VM, we need to emulate HIE/HIP and HGEIE/HGEIP for the host VM
so that it can figure out which guest TVM has a pending external interrupt.

For HIE/HIP, we only properly virtualize the SG_EXT bit as we don't intend
to support injection of VS_{SOFT,EXT,TIMER} into the host VM as it's of
limited usefulness (at least for Linux/KVM). We still let the host VM write
the bits in HIE.

For HGEIE/HGEIP, we make the assumption that the hosting VM's guest
interrupt files are immediately contiguous to it in HGEIE/HGEIP. That works
just fine for now as we don't support further nesting (i.e. a guest TVM
hosting another guest TVM). The host VN's HGEI bit is set in HGEIE when
it's running a guest TVM, otherwise it is clear. This is to support
switching back to the host VM when it recieves an external interrupt
while running a guest TVM.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>