VmCpu: Support emulation of guest external interrupts
Support injection of supervisor guest external interrupts into the host VM
using HVICTL if the host has them enabled in its virtual HIE and we detect
that it has guests with pending and enabled external interrupts.

We also make use of SGEI to trap a guest TVM if the host VM has pending
and enabled external interrupts or guest external interrupts, which we
disambiguate by checking HGEIP for the host VM's interrupt file (VGEIN)
bit.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>