Use NACL extension for registering shmem area
We're going to use the same shared memory interface as the proposed NACL
extension, so to avoid defining the same function in the TEE-Host extension,
use the NACL extension for (un)registering the shared memory area. The
layout of the scratch area in the NACL shared memory structure (NaclShmem)
will still be defined by the TEE-Host extension.

This removes the TsmSetShmem call, and defines just enough of the NACL
extension to register shared memory. We don't care about the rest of the
NACL extension currently, so leave it out for now.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>