Vm: Forward ECALL return value from host
For certain ECALLs (for now, just debug console) from guest VMs, we want to
forward the call to the host and have the host provide the return value.
Do this by taking the A0/A1 values from the guest GPRs in the shared memory
area when running the vCPU after it exits due to a forwarded ECALL, just
like we do for emulated MMIO loads.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>