Add bug fix for PMU counter logic
A recent refactoring of the PMU counter logic had introduced a latent
bug that was exposed by a code change that introduced an additional
context switch between counter configuration and start. The gist is
that on the outbound path, the code change was causing the internal
bookkeeping to revert it back to the non-configured state, meaning
that the subsequent call to start the counter would fail.

This fixes the issue by ensuring the the bookkeeping code isn't
invoked on the context switch path.