tellus: Secondary CPU support
Teach tellus to activate secondary CPUs. Add a TaskRunner facility to run code
on them. The idea is that control flow remains single-threaded locally, the
main thread merely gains the ability to synchronously execute code (in the form
of Fn() trait objects, e.g. closures) on other CPUs. The new code tries to keep
things simple, in particular the TaskRunner implementation is not suitable for
production code, but should be good enough for tellus in its role as a test
workload.

Signed-off-by: Mattias Nissler <mnissler@rivosinc.com>