tellus: Use IPIs to signal TaskRunners
Previously, the secondary harts were spinning in the TaskRunner::service loop
while waiting for a new task to appear. This is obviously suboptimal but was
deemed OK for test code. It turns out that it is pretty straightforward to
avoid: This change makes the harts perform a WFI instruction when waiting for
the next task. When a task gets submitted, TaskRunner::run sends an IPI to
un-stall the target hart so it will perform another loop iteration to discover
and execute the submitted task.

Signed-off-by: Mattias Nissler <mnissler@rivosinc.com>