riscv-pages: Implement `From<Pte>` for `PageAddr`
`PageAddr` has a `new` that assumes 4k alignment, add a similar
implementation of `From` that is infallible and returns a 4k aligned
address from a `Pfn`. Clarify that the existing `from_pte` should be
used when a specific alignment is needed.

This is useful as there are several places that currently pass
`(pfn, 4k)` to `from_pte` and unwrapping. Now those unwraps can be
eliminated.

Signed-off-by: Dylan Reid <dgreid@rivosinc.com>