tellus: Fix interrupt clearing in TaskRunner
When originally writing the code, I was unaware that the CSR.sie.sext bit is
read-only to software and was under the impression that it needs to be cleared.
This gets take care of automatically by hardware. Update the code to leave the
CSR.sie register alone, and adjust the comment to call out what's really
happening. Note that the previous code was functionally equivalent, so this
code change doesn't affect behavior (and I verified that interrupt clearing
works as expected with the change), but better to fix this inaccuracy before it
confuses others.