Fix htimedelta adjustment in stimecmp calculation.
htimedelta is a negative value and to program vstimecmp into
stimecmp we need to subtract it from vstimecmp to make it positive.

Let's say host started at time=10, htimedelta = -10. Now host sets
a timer at vstimecmp=5 (time=15). For this to be done using stimecmp
in Salus we will need to program stimecmp=15 i-e vstimecmp - htimedelta.

Signed-off-by: Rajnesh Kanwal <rkanwal@rivosinc.com>