VmInterrupts: Allow interrupt injection to new file while rebinding
When VCPU is not bound we sent those interrupts to swfile to be
later copied to hw-interrupt file for injection.

When rebinding, the `rebind_imsic_begin()` sets the state to
Rebinding but swfile is not set up for use as it's supposed to be
setup on the previous PCPU which is done by `rebind_imsic_clone()`.

Given sw_file is used only between `rebind_imsic_clone()` and
`rebind_imsic_finish()' it's possible that we miss interrupts
between `rebind_imsic_begin()` and `rebind_imsic_clone()`.

Given we have the new interrupt file wired up, we can send interrupts
directly to the new interrupt file in the Rebinding state and after
transitioning to Cloned state, we are safe to use swfile.

Signed-off-by: Rajnesh Kanwal <rkanwal@rivosinc.com>
Reported-by: Atish Patra <atishp@rivosinc.com>