page-tracking: Introduce new states for blocked pages
In case the page has been previously invalidated, we know it must have
been performed by the current TVM because a page can only be blocked by
a single TVM. And based on this information, we can safely release the
page by moving it to a Converted, Mapped or Shared state.

In case the page isn't marked as invalidated but is in a BlockedShared
state, this means that another TVM invalidated it. Since it's a shared
page and the current TVM isn't the one which blocked the page, there's
no reason to change the page state, except that we have to decrement the
reference count associated with the page.

Signed-off-by: Sebastien Boeuf <seb@rivosinc.com>