Consolidate 'spin' dependencies
We use the spin crate all over the place and haven't been consistent
about what features we select, which causes a bit of mess when building
with bazel. Add a sync crate which simply pulls in and re-exports the
spin features we use, and switch all existing users over to using the
symbols from this new crate. This will also make it easier to swap
out the underlying implementation if and when we need to do so.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>